Add useModal and use it in the Modal and Drawer.
