+++ extractedTranslations/en/translation.json
@@ -1,6 +1,5 @@
 {
   "My App": "My App",
-  "Ulbi TV App": "Ulbi TV App",
   "decrement": "decrement",
   "increment": "increment",
   "throw error": "throw error",
