+++ src/shared/ui/Drawer/Drawer.module.scss
@@ -49,3 +49,9 @@
         transform: translateY(0%);
     }
 }
+
+.isClosing {
+    .content {
+        transform: translateY(100%);
+    }
+}
