+++ src/shared/ui/Drawer/Drawer.tsx
@@ -1,6 +1,7 @@
 import { classNames, Mods } from 'shared/lib/classNames/classNames';
 import React, { memo, ReactNode } from 'react';
 import { useTheme } from 'app/providers/ThemeProvider';
+import { useModal } from 'shared/lib/hooks/useModal/useModal';
 import { Overlay } from '../Overlay/Overlay';
 import cls from './Drawer.module.scss';
 import { Portal } from '../Portal/Portal';
@@ -10,6 +11,7 @@ interface DrawerProps {
     children: ReactNode;
     isOpen?: boolean;
     onClose?: () => void;
+    lazy?: boolean;
 }
 
 export const Drawer = memo((props: DrawerProps) => {
@@ -18,17 +20,33 @@ export const Drawer = memo((props: DrawerProps) => {
         children,
         onClose,
         isOpen,
+        lazy,
     } = props;
     const { theme } = useTheme();
 
+    const {
+        close,
+        isClosing,
+        isMounted,
+    } = useModal({
+        animationDelay: 300,
+        onClose,
+        isOpen,
+    });
+
     const mods: Mods = {
         [cls.opened]: isOpen,
+        [cls.isClosing]: isClosing,
     };
+
+    if (lazy && !isMounted) {
+        return null;
+    }
 
     return (
         <Portal>
             <div className={classNames(cls.Drawer, mods, [className, theme, 'app_drawer'])}>
-                <Overlay onClick={onClose} />
+                <Overlay onClick={close} />
                 <div
                     className={cls.content}
                 >
